Refactoring. Future Calculate Polish Notation with visitor pattern.
