+++ CogPar/src/uk/co/cogitolearning/cogpar/AdditionExpressionNode.java
@@ -93,4 +93,9 @@ public class AdditionExpressionNode extends SequenceExpressionNode
       t.expression.accept(visitor);
   }
 
+  @Override
+  public void acceptOnce(ExpressionNodeVisitor visitor) {
+    visitor.visit(this);
+  }
+
 }
