+++ CogPar/src/uk/co/cogitolearning/cogpar/ConstantExpressionNode.java
@@ -83,4 +83,9 @@ public class ConstantExpressionNode implements ExpressionNode
   {
     visitor.visit(this);
   }
+
+  @Override
+  public void acceptOnce(ExpressionNodeVisitor visitor) {
+    visitor.visit(this);
+  }
 }
