+++ CogPar/src/uk/co/cogitolearning/cogpar/DivExpressionNode.java
@@ -80,4 +80,9 @@ public class DivExpressionNode extends SequenceExpressionNode
       t.expression.accept(visitor);
   }
 
+  @Override
+  public void acceptOnce(ExpressionNodeVisitor visitor) {
+    visitor.visit(this);
+  }
+
 }
